Crate interbtc_runtime_standalone

The Substrate Node Template runtime. This can be compiled with #[no_std], ready for Wasm.

Re-exports

pub use pallet_grandpa::fg_primitives;

pub use btc_relay::bitcoin;

pub use primitives;

pub use primitives::CurrencyId::Token;

Modules

api

token_distribution

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

AdjustmentVariable

The adjustment variable of the runtime. Higher values will cause TargetBlockFullness to change the fees more rapidly.

AuraAccountAdapter

BalanceWrapper

a wrapper around a balance, used in RPC to workaround a bug where using u128 in runtime-apis fails. See https://github.com/paritytech/substrate/issues/4641

BaseCallFilter

BlockExecutionWeight

Importing a block with 0 txs takes ~5 ms

BlockHashCount

BlockNumberToBalance

Burn

CurrencyConvert

DealWithFees

DealWithRewards

DustRemovalWhitelist

EmissionPeriod

EnactmentPeriod

EqualPrivilegeOnly

Implementation of [PrivilegeCmp] that only checks for equal origins.

EscrowAnnuityAccount

EscrowAnnuityPalletId

EscrowBlockRewardProvider

ExistentialDeposits

ExtrinsicBaseWeight

Executing 10,000 System remarks (no-op) txs takes ~1.26 seconds -> ~125 µs per tx

FastTrackVotingPeriod

FeeAccount

FeePalletId

GenesisConfig

GetDepositBase

GetDepositFactor

GetMaxSignatories

GetNativeCurrencyId

GetRelayChainCurrencyId

GetWrappedCurrencyId

IdentityFee

Implementor of WeightToFeePolynomial that maps one unit of weight to one unit of fee.

InflationPeriod

IssueRequest

LaunchPeriod

MaxApprovals

MaxAuthorities

MaxLocks

MaxPeriod

MaxProposals

MaxScheduledPerBlock

MaxVestingSchedules

MaxVotes

MaximumSchedulerWeight

MinVestedTransfer

MinimumDeposit

MinimumMultiplier

Minimum amount of the multiplier. This value cannot be too low. A test case should ensure that combined with AdjustmentVariable, we can recover from the minimum. See multiplier_can_grow_from_zero.

MinimumPeriod

NoPreimagePostponement

OperationalFeeMultiplier

This value increases the priority of Operational transactions by adding a “virtual tip” that’s equal to the OperationalFeeMultiplier * final_fee.

Origin

The runtime origin type representing the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParachainBlocksPerBitcoinBlock

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PreimageBaseDepositz

PreimageByteDeposit

PreimageByteDepositz

PreimageMaxSize

ProposalBond

ProposalBondMaximum

ProposalBondMinimum

RedeemRequest

RefundRequest

ReplaceRequest

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SchedulerMigrationV3

SessionKeys

Span

SpendPeriod

SupplyAccount

SupplyPalletId

TargetBlockFullness

The portion of the NORMAL_DISPATCH_RATIO that we adjust the fees with. Blocks filled less than this will decrease the weight and more will increase.

TechnicalCommitteeMaxMembers

TechnicalCommitteeMaxProposals

TechnicalCommitteeMotionDuration

TransactionByteFee

TreasuryAccount

TreasuryPalletId

UncleGenerations

VaultAnnuityAccount

VaultAnnuityPalletId

VaultBlockRewardProvider

VaultRegistryAccount

VaultRegistryPalletId

Version

VotingPeriod

Enums

Call

CurrencyId

DispatchClass

A generalized group of dispatch types.

Event

Everything

A [Contains] implementation that contains every value.

IssueEvent

The event emitted by this pallet.

NominationEvent

The event emitted by this pallet.

OriginCaller

RedeemEvent

The event emitted by this pallet.

RefundEvent

The event emitted by this pallet.

RelayCall

Contains one variant per dispatchable that can be called by an extrinsic.

RelayEvent

The event emitted by this pallet.

ReplaceEvent

The event emitted by this pallet.

StatusCode

Enum indicating the status of the BTC Parachain.

TimestampCall

Contains one variant per dispatchable that can be called by an extrinsic.

TokenSymbol

Constants

BITCOIN_BLOCK_SPACING

BITCOIN_SPACING_MS

CENTS

DAYS

DOLLARS

DOT

HOURS

INTERBTC

INTR

KINT

KSM

MILLICENTS

MILLISECS_PER_BLOCK

MINUTES

SLOT_DURATION

TARGET_SPACING

Target Spacing: 10 minutes (600 seconds)

UNITS

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WEEKS

WEIGHT_PER_SECOND

YEARS

Traits

BuildStorage

Complex storage builder stuff.

CurrencyInfo

FindAuthor

A trait for finding the author of a block header based on the PreRuntime digests contained within it.

Get

A trait for querying a single value from a type.

KeyOwnerProofSystem

Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

deposit

get_all_module_accounts

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Aura

AuraConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

BTCRelay

BTCRelayConfig

Balance

Balance of an account.

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CheckedExtrinsic

Extrinsic type that has already been checked.

Currency

Democracy

DemocracyConfig

Escrow

EscrowAnnuity

EscrowAnnuityInstance

EscrowRewards

EscrowRewardsInstance

Executive

Executive: handles dispatch to the various modules.

Fee

FeeConfig

Grandpa

GrandpaAuthorityList

A list of Grandpa authorities with associated weights.

GrandpaConfig

GrandpaId

Identity of a Grandpa authority.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Issue

IssueConfig

LockIdentifier

An identifier for a lock. Used for disambiguating different locks so that they can be individually replaced or removed.

Moment

An instant or duration in time.

MultiSig

Nomination

NominationConfig

Nonce

Index of a transaction in the chain. 32-bit should be plenty.

Oracle

OracleConfig

Preimage

Redeem

RedeemConfig

Refund

RefundConfig

Relay

Replace

ReplaceConfig

Scheduler

Security

SecurityConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedFixedPoint

The signed fixed point type.

SignedInner

The Inner type of the SignedFixedPoint.

SlowAdjustingFeeUpdate

Sudo

SudoConfig

Supply

SupplyConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalCommitteeInstance

TechnicalMembership

TechnicalMembershipConfig

Timestamp

Tokens

TokensConfig

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

UnsignedFixedPoint

The unsigned fixed point type.

UnsignedInner

The Inner type of the UnsignedFixedPoint.

Utility

VaultAnnuity

VaultAnnuityInstance

VaultRegistry

VaultRegistryConfig

VaultRewards

VaultRewardsInstance

VaultStaking

Vesting

VestingConfig

Weight

Numeric range of a transaction weight.