Trait interbtc_runtime_standalone::Get

pub trait Get<T> {
    fn get() -> T;
}

A trait for querying a single value from a type.

It is not required that the value is constant.

Required Methods

fn get() -> T

Return the current value.